@@ -1,9 +1,11 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.11'
+gem 'rails', '> 4.2.25'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
+
+gem 'annotate'
 
 group :production do
   gem 'pg'
@@ -17,14 +19,16 @@ end
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails'
+  gem 'coffee-rails'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   # gem 'therubyracer', :platforms => :ruby
 
-  gem 'uglifier', '>= 1.0.3'
+  gem 'uglifier'
 end
+
+gem 'haml-rails'
 
 gem 'jquery-rails'
 
@@ -32,14 +36,16 @@ gem 'devise'
 
 gem 'kaminari'
 
-gem 'bootstrap-sass'
+gem "twitter-bootstrap-rails"
+
+#gem 'bootstrap-sass'
 gem 'bootswatch-rails'
 
-gem 'rails_bootstrap_navbar'
+#gem 'rails_bootstrap_navbar'
 
 gem 'high_voltage'
 
-gem 'font-awesome-sass-rails'
+#gem 'font-awesome-sass-rails'
 
 gem 'carrierwave'
 gem 'rmagick'
