@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -9,60 +8,58 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130401204823) do
+ActiveRecord::Schema.define(version: 20130401204823) do
 
-  create_table "friends", :force => true do |t|
+  create_table "friends", force: :cascade do |t|
     t.string   "name"
-    t.datetime "created_at",                             :null => false
-    t.datetime "updated_at",                             :null => false
-    t.string   "email",                  :default => "", :null => false
-    t.string   "encrypted_password",     :default => "", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.string   "email",                  default: "", null: false
+    t.string   "encrypted_password",     default: "", null: false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          :default => 0
+    t.integer  "sign_in_count",          default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
     t.boolean  "is_admin"
     t.string   "avatar"
+    t.index ["email"], name: "index_friends_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_friends_on_reset_password_token", unique: true
   end
 
-  add_index "friends", ["email"], :name => "index_friends_on_email", :unique => true
-  add_index "friends", ["reset_password_token"], :name => "index_friends_on_reset_password_token", :unique => true
-
-  create_table "photos", :force => true do |t|
+  create_table "photos", force: :cascade do |t|
     t.string   "title"
     t.string   "image"
     t.string   "description"
     t.integer  "friend_id"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "tweets", :force => true do |t|
+  create_table "tweets", force: :cascade do |t|
     t.string   "message"
     t.integer  "friend_id"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
     t.integer  "photo_id"
     t.integer  "video_id"
+    t.index ["friend_id"], name: "index_tweets_on_friend_id"
   end
 
-  add_index "tweets", ["friend_id"], :name => "index_tweets_on_friend_id"
-
-  create_table "videos", :force => true do |t|
+  create_table "videos", force: :cascade do |t|
     t.string   "title"
     t.string   "image"
     t.string   "data_url"
     t.string   "source"
     t.string   "source_id"
     t.integer  "friend_id"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
 end
