@@ -1,3 +1,25 @@
+# == Schema Information
+#
+# Table name: friends
+#
+#  id                     :integer          not null, primary key
+#  name                   :string
+#  created_at             :datetime
+#  updated_at             :datetime
+#  email                  :string           default(""), not null
+#  encrypted_password     :string           default(""), not null
+#  reset_password_token   :string
+#  reset_password_sent_at :datetime
+#  remember_created_at    :datetime
+#  sign_in_count          :integer          default(0)
+#  current_sign_in_at     :datetime
+#  last_sign_in_at        :datetime
+#  current_sign_in_ip     :string
+#  last_sign_in_ip        :string
+#  is_admin               :boolean
+#  avatar                 :string
+#
+
 class Friend < ActiveRecord::Base
   # Include default devise modules. Others available are:
   # :token_authenticatable, :confirmable,
@@ -6,8 +28,8 @@ class Friend < ActiveRecord::Base
          :recoverable, :rememberable, :trackable, :validatable
 
   # Setup accessible (or protected) attributes for your model
-  attr_accessible :email, :password, :password_confirmation, :remember_me
-  attr_accessible :name, :avatar
+  # attr_accessible :email, :password, :password_confirmation, :remember_me
+  # attr_accessible :name, :avatar
   has_many :tweets
   has_many :photos
   
